Cross-compile coaxial.core: move all java.nio out to coaxial.jvm
With the four typeclass givens now delegating to the seam, evacuate the last
platform code from coaxial.core so it compiles under Scala.js:

  - the socket-option machinery (`configure`/`applyOptions`/`Configurable`) and
    interface resolution (`interfaceFor`/`bindAddress`) move to coaxial.jvm;
  - the `Duplex` java bridges (`channel`/`streams`) become `channelDuplex`/
    `streamsDuplex` helpers in coaxial.jvm; `Duplex.pair` and the trait stay in
    core;
  - `Connection`, `SecureEndpoint` and `Tls` (java streams / javax.net.ssl) move
    to coaxial.jvm, and their soundness exports move with them.

coaxial.core's dependency on `galilei.jvm` (only ever pulled in for its
transitive turbulence) is replaced with `turbulence.core`, and the `scalaJs =
false` override is dropped. `coaxial.core.js.compile` is now green; the JVM
build and all 36 tests are unchanged.
