+++ lib/coaxial/src/core/soundness_coaxial_core.scala
@@ -34,10 +34,10 @@ package soundness
 
 export
   coaxial
-  . { Bindable, BindError, Connectable, Connection, ConnectionError, Control, DomainSocket,
+  . { Bindable, BindError, Connectable, ConnectionError, Control, DomainSocket,
       DomainSocketEndpoint, duplex, Duplex, Duplexable, exchange, Ingressive, listen, Packet,
-      react, Routable, SecureEndpoint, Sender, Serviceable, SocketBackend, SocketEvent,
-      SocketOption, SocketService, Tls, Transmissible, transmit, UdpResponse }
+      react, Routable, Sender, Serviceable, SocketBackend, SocketEvent,
+      SocketOption, SocketService, Transmissible, transmit, UdpResponse }
 
 package socketOptions:
   export
