+++ lib/coaxial/src/jvm/coaxial.Connection.scala
@@ -45,7 +45,9 @@ case class Connection
   // A fresh pull endpoint over the read side; single-use, like the connection.
   def source()(using Buffering)(using tactic: Tactic[StreamError])
   :   (Stream[Data] over Credit)^{tactic, caps.any} =
+
     summon[(ji.InputStream is Source by Data over Credit)^].stream(in)
+
   def reader: ji.InputStream = in
   def writer: ji.OutputStream = out
