+++ lib/coaxial/src/jvm/soundness_coaxial_jvm.scala
@@ -32,5 +32,7 @@
                                                                                                   */
 package soundness
 
+export coaxial.{Connection, SecureEndpoint, Tls}
+
 package socketBackends:
   export coaxial.socketBackends.virtualMachine
